Say the name of the post that the flag is respawning to. Added a name attribute to post.
Added supprot for multiple posts.
Added a sequential attribute to <flag> that determines if the respawning flag randomly chooses a post to respawn to or if it cycles through posts.
Added the boolean of a specified post, which is when a net sets a new post for the flag to spawn to. If it is a specified post, it doesn't care if the original flag decleration has multiple flags. That means, if the flag respawns at a post with specifiedPost = true, respawning will remain at this post until a net changes it to a new post.

Signed-off-by: mrcookie <athertonmook@gmail.com>