Cleanup timelock module check in the start of a match
Because GameRulesMatchModule already sets the value of doDaylightCycle based on the setting of <timelock> (defaults on, so always set) or doDaylightCycle in <gamerules> (overrides <timelock> if present), there's some unnecessary steps taken in PGMListener that have been simplified here. This changes PGMListener to only check for the value of the rule in GameRulesMatchModule, consistent with how rules for fire tick are set.

Signed-off-by: BT (calcastor/mame) <43831917+calcastor@users.noreply.github.com>