fix(ModernParticles): correct spawnParticle argument list for spawnFlame, explosion
This patch corrects the argument list for FLAME/EXPLOSION data types, where the boolean was supposed to force the client to show the particle but was incorrectly passed as particle data, resulting in an exception being thrown.

This solves the issue with spawners spam-spawning items as fast as the server can.

Signed-off-by: TTtie <me@tttie.cz>