fix(platform/modern): hide the bootstrap classloader, use PluginClassLoader
This patch hides the bootstrap classloader from other plugins to ensure that 3rd party integrations load PGM classes from the remapping classloader.

Also, this patch gets rid of the PGMClassLoader in favor of PluginClassLoader, which still does the legacy remapping but is built into Paper and as such removes the maintenance burden of having to bring our own classloader.

Signed-off-by: TTtie <me@tttie.cz>