refactor!: remove FilterBuilder#respondsTo
The usage of Filter#respondsTo hints that it is meant to be used with dynamic filters, which are covered by FilterBuilder#dynamic.

This affects the behavior of the switch-case replacement by allowing any filter to be passed. This also means that filters that abstain will match if used in a case branch.

Signed-off-by: TTtie <me@tttie.cz>