fix: fix iron doors chopping off in modern
Fixes an issue where iron doors would get chopped off on modern servers due to Block#getData returning unreliable and/or contradicting data there, causing the block to be treated as an "invisible half-iron-door" and get replaced with a barrier.

Additionally, this patch exempts >=1.13 worlds (based on the data version in level.dat) from the aforementioned behavior; those worlds are not loadable on 1.8 by default and as such I think it's reasonable to assume that the map author isn't using a half door to create an invisible barrier.

Signed-off-by: TTtie <me@tttie.cz>