fix: make tool repair work better with arrows on modern
It turns out that picking up arrows needs special handling on modern, in particular, the item and arrow entities have been decoupled, so everything involving picking up the item has to be done on the arrow entity instead.

This fixes the issue with "infinite pickup" on maps that have arrows (or tridents) in tool repair.

Signed-off-by: TTtie <me@tttie.cz>